requirements.txt
imageio==2.31.1
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
isort==5.11.5
jsonschema==4.6.0
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyter-server==1.17.1
jupyterlab==3.4.3
jupyterlab-pygments==0.2.2
jupyterlab-server==2.14.0
matplotlib==3.5.3
matplotlib-inline==0.1.3
numpy==1.21.6
openai==0.25.0
opencv-python==4.7.0.68
pickleshare==0.7.5
Pillow==9.4.0
plotly==5.8.2
pydub==0.25.1
pyflakes==2.5.0
Pygments==2.10.0
pylint==2.17.1
PyMySQL==1.0.2
pyparsing==3.0.9
pypiwin32==223
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.1
pytoolconfig==1.2.5
pyttsx3==2.90
pytz==2022.1
PyWavelets==1.3.0
pywin32==302
pywinpty==2.0.5
pyzmq==22.3.0
requests==2.28.0
rope==1.7.0
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.7
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116

